update tests location
